+++ UcbBack/Models/Dist_Pregrado.cs
@@ -12,7 +12,7 @@ namespace UcbBack.Models
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public int Id { set; get; }
+        public long Id { set; get; }
 
         public string Document { get; set; }
         public string FirstName { get; set; }
@@ -32,6 +32,6 @@ namespace UcbBack.Models
         public string gestion { get; set; }
 
         public Dist_File DistFile { get; set; }
-        public int DistFileId { get; set; }
+        public long DistFileId { get; set; }
     }
 }
